Move OscoreSetting from server-core to core module.
To be reused by client.
