+++ leshan-server-core/src/main/java/org/eclipse/leshan/server/security/SecurityInfo.java
@@ -20,8 +20,8 @@ import java.io.Serializable;
 import java.security.PublicKey;
 import java.util.Arrays;
 
+import org.eclipse.leshan.core.oscore.OscoreSetting;
 import org.eclipse.leshan.core.util.Validate;
-import org.eclipse.leshan.server.security.oscore.OscoreSetting;
 
 /**
  * The security info for a client.
